+++ mycodo/inputs/si7021_circuitpython.py
@@ -1,0 +1,79 @@
+# coding=utf-8
+import copy
+
+from mycodo.inputs.base_input import AbstractInput
+
+# Measurements
+measurements_dict = {
+    0: {
+        'measurement': 'temperature',
+        'unit': 'C'
+    },
+    1: {
+        'measurement': 'humidity',
+        'unit': 'percent'
+    }
+}
+
+# Input information
+INPUT_INFORMATION = {
+    'input_name_unique': 'SI7021_CIRCUITPYTHON',
+    'input_manufacturer': 'Silicon Labs',
+    'input_name': 'Si7021',
+    'input_library': 'Adafruit-CircuitPython-Si7021',
+    'measurements_name': 'Temperature/Humidity',
+    'measurements_dict': measurements_dict,
+    'url_datasheet': 'https://www.silabs.com/documents/public/data-sheets/Si7021-A20.pdf',
+
+    'options_enabled': [
+        'i2c_location',
+        'period',
+        'pre_output'
+    ],
+    'options_disabled': ['interface'],
+
+    'dependencies_module': [
+        ('pip-pypi', 'usb.core', 'pyusb==1.1.1'),
+        ('pip-pypi', 'adafruit_extended_bus', 'Adafruit-extended-bus==1.0.1'),
+        ('pip-pypi', 'adafruit_si7021', 'adafruit-circuitpython-si7021==3.3.5')
+    ],
+
+    'interfaces': ['I2C'],
+    'i2c_location': ['0x40'],
+    'i2c_address_editable': False,
+}
+
+
+class InputModule(AbstractInput):
+    """ A sensor support class that measures """
+    def __init__(self, input_dev, testing=False):
+        super(InputModule, self).__init__(input_dev, testing=testing, name=__name__)
+
+        self.sensor = None
+
+        if not testing:
+            self.initialize_input()
+
+    def initialize_input(self):
+        import adafruit_si7021
+        from adafruit_extended_bus import ExtendedI2C
+
+        self.sensor = adafruit_si7021.SI7021(
+            ExtendedI2C(self.input_dev.i2c_bus),
+            address=int(str(self.input_dev.i2c_location), 16))
+
+    def get_measurement(self):
+        """ Gets the measurements """
+        if not self.sensor:
+            self.logger.error("Input not set up")
+            return
+
+        self.return_dict = copy.deepcopy(measurements_dict)
+
+        if self.is_enabled(0):
+            self.value_set(0, self.sensor.temperature)
+
+        if self.is_enabled(1):
+            self.value_set(1, self.sensor.relative_humidity)
+
+        return self.return_dict